Add support for the VSCode mark sequences*
\*that we already support

VsCode has their own fork of the FTCS marks. Theirs uses `633` as the OSC prefix
instead of `133`. This adds support to the Terminal for that format, as an alias
for the FTCS marks we already support.

This does not add any of the other sequences we didn't already support.

see: #11000