**BODGY** Migrate resw files from main
This works around an issue we encountered in Touchdown where we added
keys in main *and* changed keys in main and release-1.18. Because the
file hashes don't match, this file is not considered a localization
target. I think it is because we changed keys on both branches, but
there was no intermediate time point where the entire file was
consistent on both branches. Therefore, the files got localized on main
and the translations logged against the specific resw version that
existed on main. They were never logged against this weird backported
verison.