Switch the schema depending on the branding we're being built for
Ever since we started writing the entire settings file out ourselves,
we've had the opportunity to control which schema it uses.

This is a quality-of-life improvement for Preview users, and might make
life easier for Dev users as well.

One thing to consider is whether, if building Terminal in Debug+Dev
mode, we should substitute in a `file:///` path. That would make the
_local_ dev experience way cooler and cost us practically nothing.