Use a different schema for each branding (#15856)
Switch the schema depending on the branding we're being built for

Ever since we started writing the entire settings file out ourselves,
we've had the opportunity to control which schema it uses.

This is a quality-of-life improvement for Preview users, and might make
life easier for Dev users as well.

For Debug builds, it even switches over to a local `file://` path to
the schema in the source directory!

Closes #6601

(cherry picked from commit 333fcd89b305a1c4f08b05c238c39b56eb4d03e9)
Service-Card-Id: 90281511
Service-Version: 1.18