Add Auto Prop Name to 'Delete Color Scheme' button (#15994)
Since the "delete color scheme" button is filled with an icon and a Text
Box, the text is not automatically exposed as the autoProp.Name for the
button. We have to do it manually just like we do for "delete profile".

Validated manually using accessibility insights

Closes #15984

(cherry picked from commit 3550e197513368ed20a3ecc0cbcd560fe34661fb)
Service-Card-Id: 90584974
Service-Version: 1.18