Don't end the current mark, if we get one of the same kind
If you're already in the "output" state, then an app requesting an "output" mark
probably shouldn't end the current mark and start a new one. It should just keep
on keepin' on.

The decision to end the previous one was arbitrary in the first place, so let's
arbitrarily change it back.

Especially noticable if you hit <kbd>Enter</kbd> during a command, because the
auto-mark prompt work will do a CommandEnd, so long-running commands will get
broken into multiple marks 🥲