WIP
This code rejoins neighboring search results, but is that really such a
common scenario that it outweighs the cost of turning this O(n) outer
loop into O(n^2)?

Previously we only had a 1 selection overlay per row at most (the
regular, manual selection by a user). Now, with an arbitrary number of
disjoint selections per row, the technically best approach is to resize
RenderingPayload::colorBitmap to contain 3 textures and not just 2.
Right now the bitmap contains the background and foreground bitmaps back
to back, but it could also contain a bitmap for the selection. Then, in
BackendD3D we simply overlay this bitmap on top of the previously drawn
text, similar to how we draw the background as the first step.
But making this change is fairly challenging and I don't expect you to
make it.