No longer split the strings in the resource file (#16380)
Instead of splitting up the resource strings in the resource file (which
will cause localization issues), we now use resource strings with
placeholders.

Unfortunately, we still need to split the string to bind to xaml
correctly since only part of the string should be hyperlinked. We do
this in the code-behind, with the help of a helper function
`SplitResourceStringWithPlaceholders`. Reviewers should start by looking
at that function.

## Validation Steps Performed
Hyperlinked text shows up as expected